Fixed pathologic behavior at extremes.
The algorithm had allowed for arbitrarily high intensities around events that
occured at or near the endpoints of the time interval investigated, and tended to
give spiky intensity behavior there. Changed the algorithm to require a minimum set
of events per bin (default 3, counting range endpoints), which removes this
problem, and places reasonable constraints on the maximum intensity estimated
in the interior of the range.

Also, re-coded the computation to be slightly more efficient and (hopefully) a
little clearer.