Auc correction (#12)
* BREAKING: Fix intensity calculation bug.

The area under an intensity curve should be approximately equal to the number
of events that produced the curve, getting closer to exactly equal as the
number of iterations grows. But in most cases, the area was asymptotically a
tiny amount larger than the number of events, which was most obvious when there
was only one event, and therefore only one bin, and a constant intensity
regardless of the number of iterations.

The problem lied in the semantics of how the curves are described. In parts of
the code, the semantics of a curve (x, y) were that x represented bin edges,
y[i] contained the computed density of events in the bin defined by edges x[i]
and x[i+1], and len(y) = len(x) - 1. But at some point in the calculation, the
semantics changed so that x represented sampling points of the intensity curve,
so that y[i] represented the instantaneous intensity at point x[i], and we
expected len(y) == len(x). This fix resolves that inconsistency by treating x
as bin edges always. The return value from

infer_intensity(events, grid, iterations, min_count)

now returns a vector that is one element shorter than grid. This
will break code that assumes they are the same size.

* Update pxd file to refer to renamed function.

* Update tests

Co-authored-by: Tom Lasko <tom.lasko@vanderbilt.edu>